+++ src/main/java/com/marija/students/service/impl/StudentServiceImpl.java
@@ -1,13 +1,17 @@
 package com.marija.students.service.impl;
 
+import com.marija.students.dto.StudentDto;
+import com.marija.students.exception.ResourceNotFoundException;
 import com.marija.students.model.Fakultet;
 import com.marija.students.model.Student;
 import com.marija.students.repository.FakultetRepository;
 import com.marija.students.repository.StudentRepository;
 import com.marija.students.service.StudentService;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
@@ -15,8 +19,11 @@ import java.util.Optional;
 
 
 @Service
+@Transactional
 public class StudentServiceImpl implements StudentService {
 
+    @Autowired
+    private ModelMapper modelMapper;
     @Autowired
     private StudentRepository studentRepository;
 
@@ -30,15 +37,18 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student createStudent(Student student) {
+    public Student createStudent(StudentDto studentDto) {
+       Optional<Student> studentOptional = studentRepository.findStudentById(studentDto.getBrojIndeksa());
 
-        Optional<Student> studentOptional = studentRepository.findById(student.getBrojIndeksa());
-        if (studentOptional.isPresent()){
-            throw new IllegalStateException("Student postoji");
-        }else{
-            student.setStarost(calcStarost(student.getDatumRodjenja()));
-            return studentRepository.save(student);
-        }
+       if (studentOptional.isPresent()){
+           throw new IllegalStateException("Student postoji");
+       }else{
+           Student student = modelMapper.map(studentDto,Student.class);
+           student.setStarost(calcStarost(studentDto.getDatumRodjenja()));
+           Fakultet fakultet = fakultetRepository.findFakultetById(studentDto.getFakultetId());
+           student.registrujStudentaZaFakultet(fakultet);
+           return studentRepository.save(student);
+       }
     }
 
     @Override
@@ -48,7 +58,13 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void delete(String studentId) {
-        studentRepository.deleteById(studentId);
+
+        Optional<Student> student = studentRepository.findStudentById(studentId);
+        if (student.isPresent()){
+            student.get().getFakulteti().forEach(fakultet -> fakultet.getStudenti().remove(student));
+//            studentRepository.flush();
+            studentRepository.deleteById(studentId);
+        }
     }
 
     @Override
@@ -57,20 +73,24 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student assignFakultetToStudent(String studentId, String fakultetId) {
+    public Student updateStudent(StudentDto studentDto) {
 
-        List<Fakultet> listaFakulteta = null;
+        Student existingStudent = studentRepository.findStudentById(studentDto.getBrojIndeksa()).orElseThrow(
+                () -> new ResourceNotFoundException("Student ", "Broj indeksa= ", studentDto.getBrojIndeksa()));
 
-        Student student = studentRepository.findById(studentId).get();
-        Fakultet fakultet = fakultetRepository.findById(fakultetId).get();
+        existingStudent.setIme(studentDto.getIme());
+        existingStudent.setPrezime(studentDto.getPrezime());
+        existingStudent.setJmbg(studentDto.getJmbg());
+        existingStudent.setDatumRodjenja(studentDto.getDatumRodjenja());
+        existingStudent.setStarost(calcStarost(studentDto.getDatumRodjenja()));
+        existingStudent.registrujStudentaZaFakultet(fakultetRepository.findFakultetById(studentDto.getFakultetId()));
 
-        listaFakulteta = student.getAssignedFakulteti();
-        listaFakulteta.add(fakultet);
+        studentRepository.save(existingStudent);
 
-        student.setAssignedFakulteti(listaFakulteta);
+        return existingStudent;
 
-        return studentRepository.save(student);
     }
+
 
     @Override
     public int calcStarost(LocalDate datumRodjenja) {
